+++ src/main/java/com/sheepshop/entitys/Brand.java
@@ -7,9 +7,6 @@ import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.Nationalized;
 
-import java.util.LinkedHashSet;
-import java.util.Set;
-
 @Getter
 @Setter
 @Entity
@@ -18,7 +15,7 @@ public class Brand {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", nullable = false)
-    private Integer id;
+    private Long id;
 
     @Size(max = 50)
     @Nationalized
@@ -33,6 +30,5 @@ public class Brand {
     @ColumnDefault("0")
     @Column(name = "status")
     private Integer status;
-
 
 }
