+++ src/main/java/com/sheepshop/repositorys/ColorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ColorRepository extends JpaRepository<Color, Integer>{
+public interface ColorRepository extends JpaRepository<Color, Long>{
 
 }
