+++ src/main/java/com/sheepshop/services/CategoryServiceImpl.java
@@ -20,7 +20,7 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public Category findById(Integer id) {
+    public Category findById(Long id) {
         return categoryRepository.findById(id).get();
     }
