+++ src/main/java/com/sheepshop/services/RoleServiceImpl.java
@@ -30,7 +30,7 @@ public class RoleServiceImpl implements RoleService {
     }
 
     @Override
-    public Role findById(Integer id) {
+    public Role findById(Long id) {
         return roleRepository.findById(id).get();
     }
